Add script for adding LIWC entities to FoLiA XML file
Added a script that checks whether a word in the FoLiA file also occurs
in the historic LIWC dictionary. If it does, and the word has certain
categories (at this time the script checks for posemo and negemo
words), an entity is added to the FoLiA XML file.
What needs to be added is: a function that generates an appropriate id
for entities, an entity-annotation tag, and a way to customize at run
time what entities will be added.