Add t attribute to entities
The wref tag in FoLiA entities refers to the words that make up an
entity. The wref tag has an (optional) t attribute, that repeats the
word the tag refers to. Because with a t attribute it is easier to see
what word(s) an entity refers to, a t attribute was added for LIWC
entities.