Update write xml output to file
A function was added for writing the xml output to file. The function
uses the codecs.open function and writes to the location of the input
file, instead of to the location of the script. A string is appended to
the file name of the input FoLiA file, so the input file is not
overwritten.