Add all LIWC entities to FoLiA
Previously, entities were added to a new FoLiA file for LIWC
categories 'negemo' and 'posemo' only. The script now adds all LIWC
categories to the FoLiA file. This makes it easier to visualize other
LIWC categories that might also be interesting to explore in the
context of the Embodied Emotions project.