Fix including speaker names in kaf files
Previously, speaker names were manually added to the kaf-files, because
they were not part of the text. Because some speaker names contain
stage directions, all speaker names are now added as stage direction
before the speaker turn. This has the added benefit that speaker names
are now part of the text and adding them to kaf is trivial. Before,
speaker names were 'manually' added to kaf. This is no longer required.
Therefore, the folia2kaf script was updated to include speaker names in
the kaf file in the proper way.