Add leading and trailing zeros to smoothed results
Smoothing using a moving average leads to losing a few values at the
beginning and at the end of the array containing the results. To make
the smoothed results easier to print (to file), a function that append
the right amount of zeros to to either side of the array was added.