Add script that translates kaf to folia
Added a script that takes as input a .tag file (the format produced by
the KafAnnotator) and the FoLiA file that was used to generate the .kaf
file that was annotated and extracts the annotations and adds them to
the FoLiA file.
At the moment, found annotations are printed instead of added to the
FoLiA. Also, it is not yet clear how relations should be added.