Add return values to inspect_element
Added return values to inspect_element. The function now returns true
if there are no (completely) unexpected elements and false otherwise.
Element types that are added with a message and condition in the
not_expected input parameter are ignored. The return value was added
because the inspect_folia script will be used in a shell script for the
generation of kaf files. Kaf files will only be generated if the folia
file is valid according to the inspect_folia script.