Change folia2kaf script to accept output dir
The folia2kaf script was changed to accept an additional argument that
specifies the directory where the generated kaf-files should be saved.
This change was made, because we would like to put the kaf files for a
play in a separate directory and thus needs to be specified at runtime.