Add bash script to generate kaf files
Added a bash script that generates kaf-files for each (FoLiA XML) file
in some directory. The script takes two arguments: 1) a path to the
directory containing the FoLiA XML files, and 2) a path to the
directory where the output should be saved. FoLiA XML files are first
checked using the inpect_folia script and, if found valid, kaf files
are created using the folia2kaf script. For invalid files, the id is
saved to a list and the error report from inspect_folia is saved to a
text file named <play_id>.txt.