Fix check for known unexpected elements
There is a distinction between known unexpected elements and unknown
unexpected elements. Previously, documents were considered to be valid
if they contained known unexpected elements. However, this is not
right, because it is not clear how to translate these known unexpected
elements to kaf. Therefore, the script now judges a folia document
invalid if a known unexpected element is found.