Move assignment of all elements to be inspected to the top
There are plays without scenes. The inspect_folia script was expecting
that all plays have scenes. An act in a play without scenes is expected
to have different elements than an act in a play with scenes.
Assignment of elements to be inspected was moved to the top of the
script, so that expected elements can be made dependent on the
occurrence of other elements.