Update generate_kaf (disable checking of folia files)
The generate_kaf script was updated not to check folia files before
generating kaf files, because only one of the selected texts was valid
according to the inspect_folia script. And upon inspection of the error
reports, the problems with the folia files were only minor. Therefore it
was decided to generate the kaf files without checking the files first.