Add act2kaf function
Some folia files can't be parsed using BeautifulSoup or standard
lxml.parse(), probably because of memory issues. Therefore the folia2kaf
script will be refactored to use lxml.iterparse(). In preparation for
this change, the code that generates a kaf file for an act was moved to
a separate function. The idea is to parse the entire document with
iterparse, extract the xml for an act, load that part into memory,
and than use the existing functionality to generate a kaf file.