Add script to store the contents of a FoLiA XML file in ElasticSearch
Added a script that parses a FoLiA XML file containing a play and adding
the contents to an ElasticSearch index. The script extracts the text
(and other metadata) from events and stores each event as a separate
document. The next step is to (also) add entities to the documents.
Also, the script does not yet deal with sub events. This must be fixed.