Update folia2es script to process all files in a directory
The folia2es script was updated to accept a directory name as input
argument (instead of a file name) and process all files in the directory
(instead of the single file). In addition, processed (and unused) xml
elements are deleted, to make the script more memory efficient. The next
step is to properly deal with subevents.