Fix empty actor error
The extract_character_name function crashes when it is called with an
empty string (or None). Although this is not supposed to happen, it does
happen in some FoLiA XML files. Therefore the function was updated to
return the string 'UNKNOWN' when it is called with an empty string or None.