Fix lxml.etree.XMLSyntaxError
Running the folia2es script resulted in an 'lxml.etree.XMLSyntaxError: ID
XXX already defined' error. The solution (found by googling) is to set
the option huge_tree=True when creating the parser. The error occurs,
because lxml stores the IDs of all elements in some global table and
runs into troubles when the table becomes too big (this also explains
why this problem did not occur for generating the kaf files, because
that was done using a bash script that started python for each folia
file separately). More details about the error:
https://mailman-mail5.webfaction.com/pipermail/lxml/2014-May/007185.html