Fix adding subevents
Despite a previous fix, subevents were still added separately to the
index (and event containing subevents was added and the subevents were
added). This is now fixed: an event is only added to the index, if it is
not a subevent. This means that we assume that subevents are part of the
parent events. If a subevent contains sentences (this is not always the
case), the sentences are added to the parent event.