Added argument for output directory
The add_liwc_entities script now takes an additional argument that
specifies the directory that is used to write the output file to. This
additional argument will be used by the shell script for batch adding
LIWC entities to FoLiA files. The output files are saved to a different
directory so that the original files are not changed and do not get
mixed up with the FoLiA that contain the LIWC entities.