Fix folia2text for large number of folia files
The folia2text script crashed when run for a large number of folia files
(>33). The error message was the same for when a very large (or
mulitple) xml file(s) without the option huge_tree=True. However, this
option was set correctly, and the error still occurred. To fix this
error, the script was rewritten to process a single folia file (instead
of for all files in a directory). Also a bash script was added to run
folia2text for all files in a directory.