Add script to generate pairs of LIWC categories
Added a script that generates pairs of LIWC categories for Body and
Posemo words. The pairs are added to the index. And data about the pairs
is written to std out.
The LIWC categories for which pairs are generated are hardcoded. This
should be changed. Also adding the pairs to the ES index should be made
optional.