Add script to create a historic->modern dictionary
The script takes as input a modern->historic variants dictionary in a
json file (created by the generate_historic_liwc script) and
reverses it to a dictionary that can be used to replace historic words
with their modern variant(s). Currently, the script only prints
statistics about the reversal (e.g., the number of modern and historic
words, frequencies of historic words that map to multiple modern words).