Add a script that fixes tag files
The kaf files used for annotation were generated using folia files
contain strange ids (starting with lg-added). Unfortunately, these
folia files were not saved. Therefore, the word ids in the kaf files
must be updated to the word ids in the new folia files. This script does
that for hardcoded input.
The next step is to turn these parameters into command line arguments
and to call this script with a bash script that runs the fix_tags script
for all tag files in a directory.