Update folia2dataset_emo_sentences script to use folia tokenization
The folia2dataset_emo_sentences script was updated to use the folia
tokenization and print all folia words with a space between them for
each sentence. This makes tokenization easier for scikit learn (just
split on whitespace) and it makes it easier to create a spelling
normalized version of the data.