Add script to merge data and labels from different files
The script takes the sentences from one file and merges them with the
labels from another file. This script is used to generate files
containing predicted heem labels for sentences without spelling
normalization. This unnormalized data is used to expand body parts
(based on the prediction of the label 'Lichaamsdeel' and the occurrence
 of certain words in the sentence).