Add script that generates a naf file from a folia file
The difference between this script and the folia2kaf script is that the
folia2kaf script generates a file for each act, while the folia2naf
script generates a single naf file for a folia file. Most ids are the
same. However, because the term ids started at 1 for each act, the term
ids do differ between the kaf and the naf.